moved functions which compute blurred image / visibilities to calc_image module
